+++ crates/eips/src/lib.rs
@@ -3,6 +3,7 @@
     html_logo_url = "https://raw.githubusercontent.com/alloy-rs/core/main/assets/alloy.jpg",
     html_favicon_url = "https://raw.githubusercontent.com/alloy-rs/core/main/assets/favicon.ico"
 )]
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(feature = "std"), no_std)]
 
@@ -26,10 +27,10 @@ pub mod eip2935;
 
 pub mod eip4788;
 
-pub mod eip4895;
-
 pub mod eip4844;
 pub use eip4844::{calc_blob_gasprice, calc_excess_blob_gas};
+
+pub mod eip4895;
 
 pub mod eip6110;
 pub mod merge;
@@ -39,3 +40,5 @@ pub mod eip7002;
 pub mod eip7251;
 
 pub mod eip7685;
+
+pub mod eip7702;