+++ crates/rpc-types-eth/src/transaction/request.rs
@@ -364,10 +364,8 @@ impl TransactionRequest {
     ///
     /// # Returns
     ///
-    /// - Ok(type) if all necessary keys are present to build the preferred
-    /// type.
-    /// - Err((type, missing)) if some keys are missing to build the preferred
-    /// type.
+    /// - Ok(type) if all necessary keys are present to build the preferred type.
+    /// - Err((type, missing)) if some keys are missing to build the preferred type.
     pub fn missing_keys(&self) -> Result<TxType, (TxType, Vec<&'static str>)> {
         let pref = self.preferred_type();
         if let Err(missing) = match pref {
@@ -505,9 +503,21 @@ impl TransactionInput {
         Self::maybe_input(Some(data))
     }
 
+    /// Creates a new instance with the given input data and sets both `input` and `data` fields to
+    /// the same value.
+    pub fn both(data: Bytes) -> Self {
+        Self::maybe_both(Some(data))
+    }
+
     /// Creates a new instance with the given input data.
     pub const fn maybe_input(input: Option<Bytes>) -> Self {
         Self { input, data: None }
+    }
+
+    /// Creates a new instance with the given input data and sets both `input` and `data` fields to
+    /// the same value.
+    pub fn maybe_both(input: Option<Bytes>) -> Self {
+        Self { data: input.clone(), input }
     }
 
     /// Consumes the type and returns the optional input data.