+++ crates/serde/src/lib.rs
@@ -18,6 +18,9 @@ use serde::Serializer;
 mod bool;
 pub use self::bool::*;
 
+mod optional;
+pub use self::optional::*;
+
 #[cfg_attr(not(test), deprecated = "use `quantity::{self, opt, vec}` instead")]
 pub mod num;
 #[allow(deprecated)]